1. add, delete task methods 2. template tag to work with json 'tags.py' 3. tasks css 4. tasks js to show popoUp and completed tasks

+++ src/todo/services/todo_task.py
@@ -19,7 +19,7 @@ def get_tasks(category_slug: str, user_id: int):
 
 
 def get_all_tasks(user_id: int):
-    tasks = Task.objects.filter(user_id=user_id).order_by('-is_completed',
+    tasks = Task.objects.filter(user_id=user_id).order_by('is_completed',
                                                           '-date_created')
     return tasks
 
@@ -27,7 +27,7 @@ def get_all_tasks(user_id: int):
 def get_important_tasks(user_id: int):
     tasks = Task.objects.filter(
         user_id=user_id,
-        is_important=True).order_by('-is_completed', '-date_created')
+        is_important=True).order_by('is_completed', '-date_created')
 
     return tasks
 
@@ -36,7 +36,7 @@ def get_my_day_tasks(user_id: int):
     today = date.today()
     tasks = Task.objects.filter(
         user_id=user_id,
-        date_created__date=today).order_by('-is_completed', '-date_created')
+        date_created__date=today).order_by('is_completed', '-date_created')
 
     return tasks
 
@@ -46,7 +46,7 @@ def get_custom_category_tasks(category_slug: str, user_id: int):
     tasks = Task.objects.filter(
         user=User.objects.get(id=user_id),
         category=Category.objects.get(slug=category_slug)).order_by(
-            '-is_completed',
+            'is_completed',
             '-date_created')
 
     return tasks
@@ -62,7 +62,8 @@ def add_task(text: str, category_slug: str, user_id: int):
     task = Task.objects.create(task=text, is_important=is_important, user=user)
 
     category = Category.objects.filter(slug=category_slug, user=user)
-    if category:
+
+    if category is not None:
         task.category.add(*category)
 
     return task
@@ -71,5 +72,16 @@ def add_task(text: str, category_slug: str, user_id: int):
 def delete_task(task_id):
     """Delete task by id"""
     task = Task.objects.get(id=task_id)
+    if task is None:
+        return
+
     task.delete()
 
+def finish_task(task_id):
+    task = Task.objects.get(id=task_id)
+    if task is None:
+        return
+
+    if task.is_completed is False:
+        task.is_completed = True
+        task.save()
